Made some changes and added ObjectSet
A .gitignore file has been added to exclude the file to make the build.
A minified version has been added.
ObjectSet has been added, with it it's possible to render sets of CanvasObjects instead of rendering one by one. Also, if one or more objects aren't rendered for more than 2 `Viewer.prototype.clear()` calls it'll not be clickable/touchable anymore.