Some fixes and additions to TextBox.
The property "direction" has been added to TextBox.
`TextBox.prototype.getTrueXPos()` can get the position of textAlign to make it more precise. This piece of code was broken and mouse/touch events were not working properly prior to this commit.
Added `ctx.direction = this.direction` to `TextBox.prototype.render()` so it can change the direction of the text if "align" is set to "start"/"end".